Users CRUD is workable now - CRUD functionalities are finished - some translation fixes - pagination fix for multiple usage

+++ src/Controller/UsersController.php
@@ -2,25 +2,41 @@
 
 namespace App\Controller;
 
+use App\Controller\Prototype\BaseController;
 use App\Entity\Users;
 use App\Form\UsersType;
 use App\Repository\UsersRepository;
-use App\Controller\Prototype\BaseController;
+use App\Service\PaginationService;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Twig\Environment;
 
 #[Route('/users')]
 class UsersController extends BaseController
 {
     /**/
-    #[Route('/', name: 'app_users_index', methods: ['GET'])]
-    public function index(UsersRepository $usersRepository): Response
+    #[Route('/{page}', requirements: ['page' => '\d+'], name: 'app_users_index', methods: ['GET'])]
+    public function index(Environment $twig, UsersRepository $usersRepository, int $page = 1): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
+        $limit = 2;
+
+        $paginateServ = new PaginationService($twig);
+
+        $paginateServ->initPagination(
+            'app_users_index',
+            $page,
+            $usersRepository->count([]),
+            $limit
+        );
+
         return $this->renderer('users/index.html.twig', [
-            'users' => $usersRepository->findAll(),
+            'users' => $usersRepository->findBy([], null, $paginateServ->getLimit(), $paginateServ->getOffset($page)),
+            'pagi' => $paginateServ->renderPagi()
         ]);
     }
 
@@ -28,6 +44,8 @@ class UsersController extends BaseController
     #[Route('/new', name: 'app_users_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $user = new Users();
         $form = $this->createForm(UsersType::class, $user);
         $form->handleRequest($request);
@@ -39,7 +57,6 @@ class UsersController extends BaseController
             return $this->redirectToRoute('app_users_index', [], Response::HTTP_SEE_OTHER);
         }
 
-//dd($form);
         return $this->renderer('users/new.html.twig', [
             'user' => $user,
             'form' => $form,
@@ -47,9 +64,11 @@ class UsersController extends BaseController
     }
 
     /**/
-    #[Route('/{id}', name: 'app_users_show', methods: ['GET'])]
+    #[Route('/show/{id}', name: 'app_users_show', methods: ['GET'])]
     public function show(Users $user): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         return $this->renderer('users/show.html.twig', [
             'user' => $user,
         ]);
@@ -59,6 +78,8 @@ class UsersController extends BaseController
     #[Route('/{id}/edit', name: 'app_users_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Users $user, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(UsersType::class, $user);
         $form->handleRequest($request);
 
@@ -77,6 +98,8 @@ class UsersController extends BaseController
     #[Route('/{id}', name: 'app_users_delete', methods: ['POST'])]
     public function delete(Request $request, Users $user, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADIMN');
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
